+++ Pandora/Paradigm/Inventory.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Inventory (module Exports) where
 
 import Pandora.Paradigm.Inventory.Optics as Exports
-import Pandora.Paradigm.Inventory.Storage as Exports
+import Pandora.Paradigm.Inventory.Store as Exports
 import Pandora.Paradigm.Inventory.State as Exports
 import Pandora.Paradigm.Inventory.Environment as Exports
 import Pandora.Paradigm.Inventory.Accumulator as Exports