+++ CHANGELOG.md
@@ -152,3 +152,4 @@
 * Use `UT` joint scheme for `Stack` and `Graph` data structures
 * Remove `Variant` type in favor of `Covariant` and `Contravariant` constraints in joint schemes
 * Add `Covariant` constraint on schema parameter in `Adaptable` type class
+* Rename `Storage` type to `Store`